parser cleanup
