+++ Ludown/.vscode/launch.json
@@ -13,10 +13,6 @@
             "args":[
                 "--in",
                 "c:\\repos\\botbuilder-tools\\ludown\\examples\\all.lu",
-                "-o",
-                "c:\\test",
-                "-n",
-                "allQnA"
             ]
 
         }
