+++ Ludown/lib/enums/parsecommands.js
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+const cmdEnum = {
+    luis: 'luis',
+    qna: 'qna'
+};
+
+module.exports = cmdEnum;
