+++ Ludown/lib/exports.js
@@ -1,8 +1,12 @@
-#!/usr/bin/env node
 /**
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-export const helpers = require('./enums/CLI-errors');
 export const translate = require('./translate-helpers');
 export const refresh = require('./toLU-helpers');
+export const parser = require('./parseFileContents');
+export const helpers = {
+    errorCodes: require('./enums/CLI-errors'),
+    parseCommands: require('./enums/parsecommands'),
+};
+
