+++ Ludown/lib/parser.js
@@ -9,15 +9,9 @@ const chalk = require('chalk');
 const txtfile = require('read-text-file');
 const LUISObjNameEnum = require('./enums/luisobjenum');
 const parseFileContents = require('./parseFileContents');
-const deepEqual = require('deep-equal');
 const retCode = require('./enums/CLI-errors');
 const helpers = require('./helpers');
-const helperClasses = require('./classes/hclasses');
-const qna = require('./classes/qna');
-const cmdEnum = {
-    luis: 'luis',
-    qna: 'qna'
-};
+const cmdEnum = require('./enums/parsecommands');
 const parser = {
     /**
      * Handle parsing the root file that was passed in command line args
@@ -45,9 +39,9 @@ const parser = {
         let allParsedQnAContent = allParsedContent.QnAContent;
         let finalLUISJSON, finalQnAJSON; 
         try {
-            finalLUISJSON = collateLUISFiles(allParsedLUISContent);
-            if(haveLUISContent(finalLUISJSON)) validateLUISBlob(finalLUISJSON);
-            finalQnAJSON = collateQnAFiles(allParsedQnAContent);
+            finalLUISJSON = parseFileContents.collateLUISFiles(allParsedLUISContent);
+            if(haveLUISContent(finalLUISJSON)) parseFileContents.validateLUISBlob(finalLUISJSON);
+            finalQnAJSON = parseFileContents.collateQnAFiles(allParsedQnAContent);
         } catch (err) {
             throw (err);
         }
@@ -279,17 +273,17 @@ const parseAllFiles = async function(filesToParse, log, luis_culture) {
             })
         } 
         try {
-            if(haveLUISContent(parsedContent.LUISBlob) && validateLUISBlob(parsedContent.LUISBlob)) allParsedLUISContent.push(parsedContent.LUISBlob);
+            if(haveLUISContent(parsedContent.LUISJsonStructure) && parseFileContents.validateLUISBlob(parsedContent.LUISJsonStructure)) allParsedLUISContent.push(parsedContent.LUISJsonStructure);
         } catch (err) {
             throw (err);
         }
-        allParsedQnAContent.push(parsedContent.QnABlob);
+        allParsedQnAContent.push(parsedContent.qnaJsonStructure);
         // remove this file from the list
         let parentFile = filesToParse.splice(0,1);
         let parentFilePath = path.parse(path.resolve(parentFile[0])).dir;
         // add additional files to parse to the list
-        if(parsedContent.fParse.length > 0) {
-            parsedContent.fParse.forEach(function(file) {
+        if(parsedContent.additionalFilesToParse.length > 0) {
+            parsedContent.additionalFilesToParse.forEach(function(file) {
                 if(path.isAbsolute(file)) {
                     filesToParse.push(file);
                 } else {
@@ -303,285 +297,6 @@ const parseAllFiles = async function(filesToParse, log, luis_culture) {
         QnAContent: allParsedQnAContent
     };
 }
-/**
- * Helper function to validate parsed LUISJsonblob
- * @param {Object} LUISJSONBlob input LUIS Json blob
- * @param {Object} entitiesList list of entities in collated models
- * @returns {Boolean} True if validation succeeds.
- * @throws {object} Throws on errors. Object includes errCode and text. 
- */
-const validateLUISBlob = function(LUISJSONBlob) {
-    // patterns can have references to any other entity types. 
-    // So if there is a pattern.any entity that is also defined as another type, remove the pattern.any entity
-    let spliceList = [];
-    if(LUISJSONBlob.patternAnyEntities.length > 0) {
-        for(let i in LUISJSONBlob.patternAnyEntities) {
-            let patternAnyEntity = LUISJSONBlob.patternAnyEntities[i];
-            if(helpers.filterMatch(LUISJSONBlob.entities, 'name', patternAnyEntity.name).length > 0) {
-                spliceList.push(patternAnyEntity.name);
-            }
-            if(helpers.filterMatch(LUISJSONBlob.closedLists, 'name', patternAnyEntity.name).length > 0) {
-                spliceList.push(patternAnyEntity.name);
-            }
-            if(helpers.filterMatch(LUISJSONBlob.model_features, 'name', patternAnyEntity.name).length > 0) {
-                spliceList.push(patternAnyEntity.name);
-            }
-            if(helpers.filterMatch(LUISJSONBlob.prebuiltEntities, 'name', patternAnyEntity.name).length > 0) {
-                spliceList.push(patternAnyEntity.name);
-            }
-        }
-    }
-    if(spliceList.length > 0) {
-        spliceList.forEach(function(item) {
-            for(let i in LUISJSONBlob.patternAnyEntities) {
-                if(LUISJSONBlob.patternAnyEntities[i].name === item) {
-                    LUISJSONBlob.patternAnyEntities.splice(i, 1);
-                    break;
-                }
-            }
-        })
-    }
-    
-    // look for entity name collisions - list, simple, patternAny, phraselist
-    // look for list entities labelled
-    // look for prebuilt entity labels in utterances
-    
-    let entitiesList = [];
-    let entityFound = '';
-    if(LUISJSONBlob.entities.length > 0) {
-        LUISJSONBlob.entities.forEach(function(entity) {
-            entitiesList.push(new helperClasses.validateLUISBlobEntity(entity.name,['simple']));
-        });
-    }
-    if(LUISJSONBlob.closedLists.length > 0){
-        LUISJSONBlob.closedLists.forEach(function(entity) {
-            entityFound = helpers.filterMatch(entitiesList, 'name', entity.name);
-            if(entityFound.length === 0) {
-                entitiesList.push(new helperClasses.validateLUISBlobEntity(entity.name,['list']));
-            } else {
-                entityFound[0].type.push('list');
-            }
-        });
-    }
-    if(LUISJSONBlob.patternAnyEntities.length > 0) {
-        LUISJSONBlob.patternAnyEntities.forEach(function(entity) {
-            entityFound = helpers.filterMatch(entitiesList, 'name', entity.name);
-            if(entityFound.length === 0) {
-                entitiesList.push(new helperClasses.validateLUISBlobEntity(entity.name,['patternAny']));
-            } else {
-                entityFound[0].type.push('patternAny');
-            }
-        });
-    }
-    if(LUISJSONBlob.model_features.length > 0) {
-        LUISJSONBlob.model_features.forEach(function(entity) {
-            entityFound = helpers.filterMatch(entitiesList, 'name', entity.name);
-            if(entityFound.length === 0) {
-                entitiesList.push(new helperClasses.validateLUISBlobEntity(entity.name,['phraseList']));
-            } else {
-                entityFound[0].type.push('phraseList');
-            }
-        });
-    }
-    // for each entityFound, see if there are duplicate definitions
-    entitiesList.forEach(function(entity) {
-        if(entity.type.length > 1) {
-            throw({
-                errCode: retCode.errorCode.DUPLICATE_ENTITIES, 
-                text: 'Entity "' + entity.name + '" has duplicate definitions.\r\n\t' + JSON.stringify(entity.type, 2, null)
-            })
-        }
-    });
-
-    // do we have utterances with labelled list entities or phraselist entities? 
-    if(LUISJSONBlob.utterances.length > 0) {
-        LUISJSONBlob.utterances.forEach(function(utterance) {
-            if(utterance.entities.length > 0) {
-                utterance.entities.forEach(function(entity) {
-                    let entityInList = helpers.filterMatch(entitiesList, 'name', entity.entity);
-                    if(entityInList.length > 0) {
-                        if(entityInList[0].type.includes('list')) {
-                            throw({
-                                errCode: retCode.errorCode.INVALID_INPUT, 
-                                text: 'Utterance "' + utterance.text + '", has reference to List entity type. \r\n\t' + 'You cannot have utterances with List entity type references in them'
-                            })
-                        }
-                        if(entityInList[0].type.includes('phraseList')) {
-                            throw({
-                                errCode: retCode.errorCode.INVALID_INPUT, 
-                                text: 'Utterance "' + utterance.text + '", has reference to PhraseList. \r\n\t' + 'You cannot have utterances with phraselist references in them'
-                            })
-                        }
-                    }
-                });
-            }
-        });
-    }
-    return true;
-}
-/**
- * Handle collating all QnA sections across all parsed files into one QnA collection
- *
- * @param {object} parsedBlobs Contents of all parsed file blobs
- * @returns {object} Final qna object
- */
-const collateQnAFiles = function(parsedBlobs) {
-    let FinalQnAJSON = new qna();
-    parsedBlobs.forEach(function(blob) {
-        // does this blob have URLs?
-        if(blob.urls.length > 0) {
-            // add this url if this does not already exist in finaljson
-            blob.urls.forEach(function(qnaUrl) {
-                if(!FinalQnAJSON.urls.includes(qnaUrl)) {
-                    FinalQnAJSON.urls.push(qnaUrl);
-                }
-            });
-        }
-        // does this blob have qnapairs?
-        if(blob.qnaList.length > 0) {
-            // walk through each qnaPair and add it if it does not exist
-            blob.qnaList.forEach(function(newQnAItem) {
-                if(FinalQnAJSON.qnaList.length == 0) {
-                    FinalQnAJSON.qnaList.push(newQnAItem);
-                } else {
-                    let qnaExists = false;
-                    let fIndex = 0;
-                    for(fIndex in FinalQnAJSON.qnaList) {
-                        if(deepEqual(FinalQnAJSON.qnaList[fIndex], newQnAItem)) {
-                            qnaExists = true;
-                            break;
-                        }
-                    }
-                    if(!qnaExists) FinalQnAJSON.qnaList.push(newQnAItem);
-                }
-            });
-        }
-    });
-    return FinalQnAJSON;
-};
-/**
- * Handle collating all LUIS sections across all parsed files into one LUIS collection
- *
- * @param {object} parsedBlobs Contents of all parsed file blobs
- * @returns {FinalLUISJSON} Final qna json contents
- * @throws {object} Throws on errors. Object includes errCode and text. 
- */
-const collateLUISFiles = function(parsedBlobs) {
-    let FinalLUISJSON = parsedBlobs[0];
-    parsedBlobs.splice(0,1);
-    parsedBlobs.forEach(function(blob) {
-        mergeResults(blob, FinalLUISJSON, LUISObjNameEnum.INTENT);
-        mergeResults(blob, FinalLUISJSON, LUISObjNameEnum.ENTITIES);
-        mergeResults_closedlists(blob, FinalLUISJSON, LUISObjNameEnum.CLOSEDLISTS);
-        mergeResults(blob, FinalLUISJSON, LUISObjNameEnum.UTTERANCE);
-        mergeResults(blob, FinalLUISJSON, LUISObjNameEnum.PATTERNS);
-        mergeResults(blob, FinalLUISJSON, LUISObjNameEnum.PATTERNANYENTITY);
-        // do we have prebuiltEntities here?
-        if(blob.prebuiltEntities.length > 0) {
-            blob.prebuiltEntities.forEach(function(prebuiltEntity){
-                let prebuiltTypeExists = false;
-                for(fIndex in FinalLUISJSON.prebuiltEntities) {
-                    if(prebuiltEntity.type === FinalLUISJSON.prebuiltEntities[fIndex].type) {
-                        // do we have all the roles? if not, merge the roles
-                        prebuiltEntity.roles.forEach(function(role) {
-                            if(!FinalLUISJSON.prebuiltEntities[fIndex].roles.includes(role)) {
-                                FinalLUISJSON.prebuiltEntities[fIndex].roles.push(role);
-                            }
-                        });
-                        prebuiltTypeExists = true;
-                        break;
-                    }
-                }
-                if(!prebuiltTypeExists) {
-                    FinalLUISJSON.prebuiltEntities.push(prebuiltEntity);
-                }
-            });
-        }
-        // do we have model_features?
-        if(blob.model_features.length > 0) {
-            blob.model_features.forEach(function(modelFeature) {
-                let modelFeatureInMaster = helpers.filterMatch(FinalLUISJSON.model_features, 'name', modelFeature.name);
-                if(modelFeatureInMaster.length === 0){
-                    FinalLUISJSON.model_features.push(modelFeature);
-                } else {
-                    if(modelFeatureInMaster[0].mode !== modelFeature.mode) {
-                        // error.
-                        throw({
-                            errCode: retCode.errorCode.INVALID_INPUT, 
-                            text: '[ERROR]: Phrase list : "' + modelFeature.name + '" has conflicting definitions. One marked interchangeable and another not interchangeable'
-                        });
-                    } else {
-                        modelFeature.words.split(',').forEach(function(word) {
-                            if(!modelFeatureInMaster[0].words.includes(word)) modelFeatureInMaster[0].words += "," + word;
-                        })
-                    }
-                }
-            });
-        }
-    }); 
-    return FinalLUISJSON;
-};
-/**
- * Helper function to merge item if it does not already exist
- *
- * @param {object} blob Contents of all parsed file blobs
- * @param {object} finalCollection Reference to the final collection of items
- * @param {LUISObjNameEnum} type Enum type of possible LUIS object types
- * @returns {void} Nothing
- */
-const mergeResults = function(blob, finalCollection, type) {
-    if(blob[type].length > 0) {
-        blob[type].forEach(function(blobItem) {
-            if(finalCollection[type].length === 0) {
-                finalCollection[type].push(blobItem);
-                return;
-            }
-            // add if this item if it does not already exist in final collection
-            let itemExists = false;
-            for(fIndex in finalCollection[type]) {
-                if(deepEqual(finalCollection[type][fIndex],blobItem)){
-                    itemExists = true;
-                    break;
-                }
-            }
-            if(!itemExists) {
-                finalCollection[type].push(blobItem);
-            }
-        });
-    }
-};
-/**
- * Helper function to merge closed list item if it does not already exist
- *
- * @param {object} blob Contents of all parsed file blobs
- * @param {object} finalCollection Reference to the final collection of items
- * @param {LUISObjNameEnum} type Enum type of possible LUIS object types
- * @returns {void} nothing
- */
-const mergeResults_closedlists = function(blob, finalCollection, type) {
-    if(blob[type].length > 0) {
-        blob[type].forEach(function(blobItem) {
-            let listInFinal = helpers.filterMatch(finalCollection[type], 'name', blobItem.name);
-            if(listInFinal.length === 0) {
-                finalCollection[type].push(blobItem);
-            } else {
-                blobItem.subLists.forEach(function(blobSLItem) {
-                    // see if there is a sublist match in listInFinal
-                    let slInFinal = helpers.filterMatch(listInFinal[0].subLists, 'canonicalForm', blobSLItem.canonicalForm);
-                    if(slInFinal.length === 0) {
-                        listInFinal[0].subLists.push(blobSLItem);
-                    } else {
-                        // there is a canonical form match. See if the values all exist
-                        blobSLItem.list.forEach(function(listItem) {
-                            if(!slInFinal[0].list.includes(listItem)) slInFinal[0].list.push(listItem);
-                        })
-                    }
-                });
-            }
-        });
-    }
-};
-
 /**
  * Helper function to see if we have any luis content in the blob
  *
